Added --chdir to start-stop-daemon
I needed to set the directory root of the node application for relevant file paths in various includes. 

There is a switch on start-stop-daemon that provides this capability. Other "how-to's" suggested setting `chdir /path/to/app` in the script part of the upstart conf but it was ignored by `start-stop-daemon`.